Conflicts:
	cocos/editor-support/cocostudio/CCArmature.cpp
	cocos/editor-support/cocostudio/CCArmature.h
	cocos/editor-support/cocostudio/CCBatchNode.cpp
	cocos/editor-support/cocostudio/CCColliderDetector.cpp
	cocos/scripting/auto-generated
	tools/bindings-generator

+++ cocos/2d/CCTileMapAtlas.cpp
@@ -222,7 +222,7 @@ void TileMapAtlas::updateAtlasValueAt(const Point& pos, const Color3B& value, in
     quad->bl.colors = color;
 
     _textureAtlas->setDirty(true);
-    int totalQuads = _textureAtlas->getTotalQuads();
+    ssize_t totalQuads = _textureAtlas->getTotalQuads();
     if (index + 1 > totalQuads) {
         _textureAtlas->increaseTotalQuadsWith(index + 1 - totalQuads);
     }
