+++ extensions/GUI/CCScrollView/CCTableView.h
@@ -105,7 +105,7 @@ public:
      * @param idx the index of a cell to get a size
      * @return size of a cell at given index
      */
-    virtual Size tableCellSizeForIndex(TableView *table, long idx) {
+    virtual Size tableCellSizeForIndex(TableView *table, ssize_t idx) {
         return cellSizeForTable(table);
     };
     /**
@@ -123,7 +123,7 @@ public:
      * @param idx index to search for a cell
      * @return cell found at idx
      */
-    virtual TableViewCell* tableCellAtIndex(TableView *table, long idx) = 0;
+    virtual TableViewCell* tableCellAtIndex(TableView *table, ssize_t idx) = 0;
     /**
      * Returns number of cells in a given table view.
      *
@@ -228,19 +228,19 @@ public:
      *
      * @param idx index to find a cell
      */
-    void updateCellAtIndex(long idx);
+    void updateCellAtIndex(ssize_t idx);
     /**
      * Inserts a new cell at a given index
      *
      * @param idx location to insert
      */
-    void insertCellAtIndex(long idx);
+    void insertCellAtIndex(ssize_t idx);
     /**
      * Removes a cell at a given index
      *
      * @param idx index to find a cell
      */
-    void removeCellAtIndex(long idx);
+    void removeCellAtIndex(ssize_t idx);
     /**
      * reloads data from data source.  the view will be refreshed.
      */
@@ -258,7 +258,7 @@ public:
      * @param idx index
      * @return a cell at a given index
      */
-    TableViewCell *cellAtIndex(long idx);
+    TableViewCell *cellAtIndex(ssize_t idx);
 
     // Overrides
     virtual void scrollViewDidScroll(ScrollView* view) override;
@@ -271,11 +271,11 @@ public:
 protected:
     long __indexFromOffset(Point offset);
     long _indexFromOffset(Point offset);
-    Point __offsetFromIndex(long index);
-    Point _offsetFromIndex(long index);
+    Point __offsetFromIndex(ssize_t index);
+    Point _offsetFromIndex(ssize_t index);
 
     void _moveCellOutOfSight(TableViewCell *cell);
-    void _setIndexForCell(long index, TableViewCell *cell);
+    void _setIndexForCell(ssize_t index, TableViewCell *cell);
     void _addCellIfNecessary(TableViewCell * cell);
 
     void _updateCellPositions();
@@ -290,7 +290,7 @@ protected:
     /**
      * index set to query the indexes of the cells used.
      */
-    std::set<long>* _indices;
+    std::set<ssize_t>* _indices;
 
     /**
      * vector with all cell positions
